Some corrections int -> long

+++ src/main/kotlin/expresions/IntExpression.kt
@@ -2,7 +2,7 @@ package expresions
 
 import machine.ProgramMemory
 
-internal class IntExpression(var value: Int) : Expression {
+internal class IntExpression(var value: Long) : Expression {
 
     override fun toOpcodes() {
         ProgramMemory.push("PUSH $value")
